backend/requirements.txt
beautifulsoup4==4.11.1
billiard==3.6.4.0
black==22.8.0
bleach==4.1.0
brotlipy==0.7.0
celery==5.2.7
config==0.5.1
debugpy==1.5.1
decorator==5.1.1
dj-rest-auth==3.0.0
Django==4.0.0
django-allauth==0.52.0
django-bootstrap4==22.2
django-cors-headers==3.14.0
django-debug-toolbar==3.7.0
django-extensions==3.2.1
django-pydenticon==0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docker==6.0.0
fastjsonschema==2.16.2
Flask==2.2.2
gunicorn==20.1.0
json5==0.9.6
jsonschema==4.4.0
multitasking==0.0.11
mypy-extensions==0.4.3
oauthlib==3.2.2
openpyxl==3.1.2
parso==0.8.3
Pillow==9.2.0
pip==22.1.2
PyMySQL==1.0.2
redis==4.3.5
requests==2.28.1
requests-oauthlib==1.3.1
soupsieve==2.3.1
SQLAlchemy==2.0.6
sqlparse==0.4.3
tqdm==4.65.0
mysqlclient==2.1.1
django_apscheduler